chore: add issues blankslate and project create from dropdown
