+++ apps/web/src/routes/_authenticated/-components/project-breadcrumb-segment.tsx
@@ -2,15 +2,17 @@ import { DropdownMenu, Text } from "@repo/ui"
 import { extractLeadingEmoji } from "@repo/utils"
 import { eq } from "@tanstack/react-db"
 import { useParams } from "@tanstack/react-router"
-import { ChevronsUpDown } from "lucide-react"
+import { ChevronsUpDown, PlusIcon } from "lucide-react"
+import { useState } from "react"
 import { useProjectsCollection } from "../../../domains/projects/projects.collection.ts"
-import { BreadcrumbText } from "./breadcrumb-ui.tsx"
+import { CreateProjectModal } from "./create-project-modal.tsx"
 
 /**
  * Project switcher / label for the header breadcrumb. Registered on `projects/$projectSlug`.
  */
 export function ProjectBreadcrumbSegment() {
   const { projectSlug } = useParams({ strict: false })
+  const [createOpen, setCreateOpen] = useState(false)
 
   const { data: project } = useProjectsCollection(
     (projects) => projects.where(({ project: p }) => eq(p.slug, projectSlug ?? "\u0000")).findOne(),
@@ -18,39 +20,41 @@ export function ProjectBreadcrumbSegment() {
   )
 
   const { data: allProjects } = useProjectsCollection()
-  const hasMultipleProjects = (allProjects?.length ?? 0) > 1
 
   if (!project || !projectSlug) return null
 
   const [emoji, title] = extractLeadingEmoji(project.name)
 
-  return hasMultipleProjects ? (
-    <DropdownMenu
-      side="bottom"
-      align="start"
-      options={
-        allProjects?.map((p) => ({
-          label: p.name,
-          onClick: () => {
-            window.location.href = `/projects/${p.slug}`
+  return (
+    <>
+      <DropdownMenu
+        side="bottom"
+        align="start"
+        options={[
+          ...(allProjects?.map((p) => ({
+            label: p.name,
+            onClick: () => {
+              window.location.href = `/projects/${p.slug}`
+            },
+          })) ?? []),
+          {
+            label: "New project",
+            iconProps: { icon: PlusIcon, size: "sm" },
+            onClick: () => setCreateOpen(true),
           },
-        })) ?? []
-      }
-      trigger={() => (
-        <button
-          type="button"
-          className="flex items-center gap-1 px-2 py-1 rounded hover:bg-muted transition-colors cursor-pointer"
-        >
-          {emoji && <span className="text-sm">{emoji}</span>}
-          <Text.H5M color="foregroundMuted">{title}</Text.H5M>
-          <ChevronsUpDown className="h-4 w-4 text-muted-foreground" />
-        </button>
-      )}
-    />
-  ) : (
-    <BreadcrumbText variant="muted">
-      {emoji && `${emoji} `}
-      {title}
-    </BreadcrumbText>
+        ]}
+        trigger={() => (
+          <button
+            type="button"
+            className="flex items-center gap-1 px-2 py-1 rounded hover:bg-muted transition-colors cursor-pointer"
+          >
+            {emoji && <span className="text-sm">{emoji}</span>}
+            <Text.H5M color="foregroundMuted">{title}</Text.H5M>
+            <ChevronsUpDown className="h-4 w-4 text-muted-foreground" />
+          </button>
+        )}
+      />
+      <CreateProjectModal open={createOpen} onClose={() => setCreateOpen(false)} />
+    </>
   )
 }
